requirements.txt
# Backend
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
networkx==3.2.1
numpy==1.26.3
scikit-learn==1.4.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.4
